+++ sdk/lib/_internal/vm/lib/typed_data_patch.dart
@@ -4020,6 +4020,8 @@ final class _Int32x4 implements Int32x4 {
       Int32x4(x ^ other.x, y ^ other.y, z ^ other.z, w ^ other.w);
   @pragma("vm:recognized", "other")
   Int32x4 operator ~() => Int32x4(~x, ~y, ~z, ~w);
+  Int32x4 andNot(Int32x4 other) =>
+      Int32x4(x & ~other.x, y & ~other.y, z & ~other.z, w & ~other.w);
   @pragma("vm:recognized", "graph-intrinsic")
   Int32x4 operator +(Int32x4 other) =>
       Int32x4(x + other.x, y + other.y, z + other.z, w + other.w);