+++ sdk/lib/typed_data/typed_data.dart
@@ -4385,6 +4385,14 @@ abstract final class Int32x4 {
   @Since("3.14")
   Int32x4 operator ~();
 
+  /// The bit-wise and-not operator.
+  ///
+  /// Returns a new [Int32x4] where each lane is the bit-wise and of the
+  /// corresponding lane of this value and the inverted corresponding lane
+  /// of [other], as if by `this & ~other`.
+  @Since("3.14")
+  Int32x4 andNot(Int32x4 other);
+
   /// Addition operator.
   Int32x4 operator +(Int32x4 other);